sqli vul2

+++ BookStoreManagement/src/main/java/com/labia/bookstoremanagement/controller/BookController.java
@@ -77,7 +77,7 @@ public class BookController {
         return bookRepository.findAll();
     }
 
-    @GetMapping("by-id")
+    @GetMapping("by-id/{bookId}")
     Book getBookById(@PathVariable("bookId") Integer id, HttpServletRequest request) {
         return bookRepository.findByBookId(id);
     }
@@ -236,9 +236,16 @@ public class BookController {
     }
 
     @DeleteMapping("delete/{bookId}")
-    public void deleteBook(@PathVariable("bookId") int bookId) {
-        bookRepository.deleteBookCategoryByBookId(bookId);
-        bookRepository.deleteById(bookId);
+    public void deleteBook(@PathVariable("bookId") int bookId,HttpServletRequest request) {
+        User user = userRepository.findByUsername(jwtTokenUtil.getUsernameFromToken(jwtTokenFilter.getJwtFromRequest(request)));
+        for (Book b:
+             user.getBooks()) {
+            if(b.getBookId() == bookId){
+                bookRepository.deleteBookCategoryByBookId(bookId);
+                bookRepository.deleteById(bookId);
+            }
+        }
+
     }
 
     @PostMapping("approve/{bookId}")
@@ -283,8 +290,8 @@ public class BookController {
      * @return
      */
     @PostMapping("add")
-    public Book addBook(@RequestBody BookData bookData) {
-        User user = userRepository.findByUsername(bookData.createdBy);
+    public Book addBook(@RequestBody BookData bookData,HttpServletRequest request) {
+        User user = userRepository.findByUsername(jwtTokenUtil.getUsernameFromToken(jwtTokenFilter.getJwtFromRequest(request)));
         bookData.book.setCreatedBy(user);
         bookRepository.save(bookData.book);
 //        Book temp = bookRepository.findByTitle(book.getTitle());
@@ -301,10 +308,10 @@ public class BookController {
     }
 
     @PostMapping("/cover/upload")
-    public void ploadCoverFile(@RequestParam("coverPath") MultipartFile file) {
+    public void ploadCoverFile(@RequestParam("coverPath") MultipartFile file,@RequestParam("bookId") String bookId) {
         String fileExtension = getFileExtension(file.getOriginalFilename());
         if ((fileExtension.equalsIgnoreCase("jpg")) && file.getSize() < 5000000) {
-            String fileName = StringUtils.cleanPath(BookId + ".jpg");
+            String fileName = StringUtils.cleanPath(bookId + ".jpg");
             try {
                 // Save the file to the uploads directory
                 String uploadDir = System.getProperty("user.dir") + COVER_UPLOAD_DIR;
@@ -317,10 +324,10 @@ public class BookController {
     }
 
     @PostMapping("/pdf/upload")
-    public void ploadPdfFile(@RequestParam("pdfPath") MultipartFile file) {
+    public void ploadPdfFile(@RequestParam("pdfPath") MultipartFile file,@RequestParam("bookId") String bookId) {
         String fileExtension = getFileExtension(file.getOriginalFilename());
         if ((fileExtension.equalsIgnoreCase("pdf")) && file.getSize() < 5000000) {
-            String fileName = StringUtils.cleanPath(BookId + ".pdf");
+            String fileName = StringUtils.cleanPath(bookId + ".pdf");
             try {
                 // Save the file to the uploads directory
                 String uploadDir = System.getProperty("user.dir") + PDF_UPLOAD_DIR;
@@ -351,22 +358,29 @@ public class BookController {
 //        return bookRepository.findById(bookId).get();
 //    }
     @PostMapping("/update/{bookId}")
-    Book updateBookById(@PathVariable Integer bookId, @RequestBody Book updateBook) {
-        Optional<Book> book = bookRepository.findById(bookId);
-        book.get().setTitle(updateBook.getTitle());
-        book.get().setDescription(updateBook.getDescription());
-        book.get().setAuthorName(updateBook.getAuthorName());
-        if (updateBook.getCategories().isEmpty()) {
-            categoryRepository.deleteBook_Category(bookId);
-        } else {
-            if (!book.get().getCategories().equals(updateBook.getCategories())) {
-                categoryRepository.deleteBook_Category(bookId);
-                for (Category c : updateBook.getCategories()) {
-                    categoryRepository.saveBook_Category(bookId, c.getCategoryId());
+    Book updateBookById(@PathVariable Integer bookId, @RequestBody Book updateBook,HttpServletRequest request) {
+        User user = userRepository.findByUsername(jwtTokenUtil.getUsernameFromToken(jwtTokenFilter.getJwtFromRequest(request)));
+        for (Book b:
+             user.getBooks()) {
+            if(b.getBookId() == bookId){
+                Optional<Book> book = bookRepository.findById(bookId);
+                book.get().setTitle(updateBook.getTitle());
+                book.get().setDescription(updateBook.getDescription());
+                book.get().setAuthorName(updateBook.getAuthorName());
+                if (updateBook.getCategories().isEmpty()) {
+                    categoryRepository.deleteBook_Category(bookId);
+                } else {
+                    if (!book.get().getCategories().equals(updateBook.getCategories())) {
+                        categoryRepository.deleteBook_Category(bookId);
+                        for (Category c : updateBook.getCategories()) {
+                            categoryRepository.saveBook_Category(bookId, c.getCategoryId());
+                        }
+                    }
                 }
+                return bookRepository.save(book.get());
             }
         }
-        return bookRepository.save(book.get());
+       return null;
     }
 
 //    @GetMapping("/{bookId}")
